+++ packages/sdk/server-ai/src/api/config/LDAIConfig.ts
@@ -1,4 +1,5 @@
 import { LDAIConfigTracker } from './LDAIConfigTracker';
+import { VercelAISDKConfig, VercelAISDKMapOptions, VercelAISDKProvider } from './VercelAISDK';
 
 /**
  * Configuration related to the model.
@@ -22,7 +23,7 @@ export interface LDModelConfig {
 
 export interface LDProviderConfig {
   /**
-   * The ID of the provider.
+   * The name of the provider.
    */
   name: string;
 }
@@ -68,13 +69,29 @@ export interface LDAIConfig {
    * Whether the configuration is enabled.
    */
   enabled: boolean;
+
+  /**
+   * Maps this AI config to a format usable direcly in Vercel AI SDK generateText()
+   * and streamText() methods.
+   *
+   * WARNING: this method can throw an exception if a Vercel AI SDK model cannot be determined.
+   *
+   * @param provider A Vercel AI SDK Provider or a map of provider names to Vercel AI SDK Providers.
+   * @param options Optional mapping options.
+   * @returns A configuration directly usable in Vercel AI SDK generateText() and streamText()
+   * @throws {Error} if a Vercel AI SDK model cannot be determined from the given provider parameter.
+   */
+  toVercelAISDK: <TMod>(
+    provider: VercelAISDKProvider<TMod> | Record<string, VercelAISDKProvider<TMod>>,
+    options?: VercelAISDKMapOptions | undefined,
+  ) => VercelAISDKConfig<TMod>;
 }
 
 /**
  * Default value for a `modelConfig`. This is the same as the LDAIConfig, but it does not include
- * a tracker and `enabled` is optional.
+ * a tracker or mapper, and `enabled` is optional.
  */
-export type LDAIDefaults = Omit<LDAIConfig, 'tracker' | 'enabled'> & {
+export type LDAIDefaults = Omit<LDAIConfig, 'tracker' | 'enabled' | 'toVercelAISDK'> & {
   /**
    * Whether the configuration is enabled.
    *