+++ packages/store/node-server-sdk-dynamodb/CHANGELOG.md
@@ -90,6 +90,17 @@
   * devDependencies
     * @launchdarkly/node-server-sdk bumped from 9.2.1 to 9.2.2
 
+## [6.2.13](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.2.12...node-server-sdk-dynamodb-v6.2.13) (2025-07-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/node-server-sdk bumped from 9.10.0 to 9.10.1
+  * peerDependencies
+    * @launchdarkly/node-server-sdk bumped from >=9.4.3 to >=9.10.1
+
 ## [6.2.12](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.2.11...node-server-sdk-dynamodb-v6.2.12) (2025-06-17)